♻️ Refactor `projwfc.x` parser
The `ProjwfcParser` relied on passing a single `out_info_dict` dictionary between
several functions to store the parsed information. This made the parser code so
entangled that it was almost unreadable and difficult to trace down bugs.

Here we refactor the class to properly isolate the various parsing functions as class
methods, and make the code more readable in general. The parent class is also changed to
the `BaseParser` one, so it can rely on the basic stdout parsing this class provides.

Another substantial change is that the `.pdos` files, which can be many and are parsed
completely, are now only temporarily retrieved instead of stored in the AiiDA
repository in the `retrieved` folder.